relacionamentos e consultas.

+++ app/src/main/java/com/example/controlenotasfrequencia/cadastroTurma/dao/TurmaDAO.java
@@ -17,7 +17,7 @@ public class TurmaDAO {
         }
     }
 
-    public static Turma getTurma(int id) {
+    public static Turma getTurma(Long id) {
         try {
             return Turma.findById(Turma.class, id);
         } catch (Exception ex) {
