+++ app/src/main/java/com/example/controlenotasfrequencia/cadastroaluno/dao/AlunoDAO.java
@@ -19,7 +19,7 @@ public class AlunoDAO {
         }
     }
 
-    public static Aluno getAluno(int id){
+    public static Aluno getAluno(Long id){
          try{
             return Aluno.findById(Aluno.class, id);
          }catch (Exception ex){
