+++ app/src/main/java/com/example/controlenotasfrequencia/domain/AlunoDisciplina.java
@@ -4,22 +4,22 @@ import com.orm.SugarRecord;
 
 public class AlunoDisciplina extends SugarRecord {
 
-    private int idAluno;
-    private int idDisciplina;
+    private Long idAluno;
+    private Long idDisciplina;
 
-    public int getIdAluno() {
+    public Long getIdAluno() {
         return idAluno;
     }
 
-    public void setIdAluno(int idAluno) {
+    public void setIdAluno(Long idAluno) {
         this.idAluno = idAluno;
     }
 
-    public int getIdDisciplina() {
+    public Long getIdDisciplina() {
         return idDisciplina;
     }
 
-    public void setIdDisciplina(int idDisciplina) {
+    public void setIdDisciplina(Long idDisciplina) {
         this.idDisciplina = idDisciplina;
     }
 }
